mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mishal Qamar", email = "mishalqamar@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
 "rich>=14.0.0",
 "opik>=1.8.42",
 "langgraph>=0.5.3",
 "langchain>=0.3.0",
 "langchain-google-genai>=2.1.8",
 "langchain-openai>=0.3.28",
 "langchain-perplexity>=0.1.1",
 "langchain-mcp-adapters>=0.1.9",
 "pydantic-settings>=2.10.0",
 "fastmcp>=0.9.0",
 "google-genai>=1.20.0",
 "python-dotenv>=1.0.0",
 "firecrawl>=4.3.6",
 "gitingest>=0.1.5",
 "httpx>=0.28.1",
 "mcp>=1.11.0",
 "nbformat>=5.10.4",
 "psutil>=7.0.0",
]

[project.scripts]
mcp-client = "mcp_client:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "mcp_client"
module-root = "src"
